+++ core/src/mast/node/loop_node.rs
@@ -32,6 +32,7 @@ impl LoopNode {
 
 /// Constructors
 impl LoopNode {
+    /// Returns a new [`LoopNode`] instantiated with the specified body node.
     pub fn new(body: MastNodeId, mast_forest: &MastForest) -> Result<Self, MastForestError> {
         if body.as_usize() >= mast_forest.nodes.len() {
             return Err(MastForestError::NodeIdOverflow(body, mast_forest.nodes.len()));
@@ -43,6 +44,12 @@ impl LoopNode {
         };
 
         Ok(Self { body, digest })
+    }
+
+    /// Returns a new [`LoopNode`] from values that are assumed to be correct.
+    /// Should only be used when the source of the inputs is trusted (e.g. deserialization).
+    pub fn new_unsafe(body: MastNodeId, digest: RpoDigest) -> Self {
+        Self { body, digest }
     }
 }