+++ processor/src/trace/tests/decoder.rs
@@ -358,10 +358,10 @@ fn decoder_p2_span_with_respan() {
 fn decoder_p2_join() {
     let mut mast_forest = MastForest::new();
 
-    let basic_block_1 = MastNode::new_basic_block(vec![Operation::Mul]);
+    let basic_block_1 = MastNode::new_basic_block(vec![Operation::Mul], None).unwrap();
     let basic_block_1_id = mast_forest.add_node(basic_block_1.clone()).unwrap();
 
-    let basic_block_2 = MastNode::new_basic_block(vec![Operation::Add]);
+    let basic_block_2 = MastNode::new_basic_block(vec![Operation::Add], None).unwrap();
     let basic_block_2_id = mast_forest.add_node(basic_block_2.clone()).unwrap();
 
     let join = MastNode::new_join(basic_block_1_id, basic_block_2_id, &mast_forest).unwrap();
@@ -423,7 +423,7 @@ fn decoder_p2_split_true() {
     // build program
     let mut mast_forest = MastForest::new();
 
-    let basic_block_1 = MastNode::new_basic_block(vec![Operation::Mul]);
+    let basic_block_1 = MastNode::new_basic_block(vec![Operation::Mul], None).unwrap();
     let basic_block_1_id = mast_forest.add_node(basic_block_1.clone()).unwrap();
 
     let basic_block_2_id = mast_forest.add_block(vec![Operation::Add], None).unwrap();
@@ -477,10 +477,10 @@ fn decoder_p2_split_false() {
     // build program
     let mut mast_forest = MastForest::new();
 
-    let basic_block_1 = MastNode::new_basic_block(vec![Operation::Mul]);
+    let basic_block_1 = MastNode::new_basic_block(vec![Operation::Mul], None).unwrap();
     let basic_block_1_id = mast_forest.add_node(basic_block_1.clone()).unwrap();
 
-    let basic_block_2 = MastNode::new_basic_block(vec![Operation::Add]);
+    let basic_block_2 = MastNode::new_basic_block(vec![Operation::Add], None).unwrap();
     let basic_block_2_id = mast_forest.add_node(basic_block_2.clone()).unwrap();
 
     let split_id = mast_forest.add_split(basic_block_1_id, basic_block_2_id).unwrap();
@@ -532,10 +532,10 @@ fn decoder_p2_loop_with_repeat() {
     // build program
     let mut mast_forest = MastForest::new();
 
-    let basic_block_1 = MastNode::new_basic_block(vec![Operation::Pad]);
+    let basic_block_1 = MastNode::new_basic_block(vec![Operation::Pad], None).unwrap();
     let basic_block_1_id = mast_forest.add_node(basic_block_1.clone()).unwrap();
 
-    let basic_block_2 = MastNode::new_basic_block(vec![Operation::Drop]);
+    let basic_block_2 = MastNode::new_basic_block(vec![Operation::Drop], None).unwrap();
     let basic_block_2_id = mast_forest.add_node(basic_block_2.clone()).unwrap();
 
     let join = MastNode::new_join(basic_block_1_id, basic_block_2_id, &mast_forest).unwrap();